Improve OpenSSL and Boost autoconfiguration.
Use more advanced autoconfiguration macros for OpenSSL and Boost.
This allows to pick up the libraries even if they are installed
in other locations than /usr (such as /usr/local on BSD) and to
specify alternative paths using the --with-openssl=/path or
--with-boost=/path switches to configure, respectively.

The downside is that autoconfiguration and build takes longer,
since the Boost macros require libtool.